fashion_news_bot/requirements.txt
requests
python-dotenv
feedparser
openai>=1.0.0
langdetect